requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-extensions==2.2.1
django-haystack==2.8.1
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
pytz==2019.2
requests==2.22.0
schedule==0.6.0
six==1.12.0
sqlparse==0.3.0
threaded==4.0.3
typing==3.7.4.1
urllib3==1.25.6
